+++ packages/@tinacms/alerts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/tinacms/tinacms/compare/@tinacms/alerts@0.1.6-alpha.0...@tinacms/alerts@0.1.6) (2020-07-07)
+
+**Note:** Version bump only for package @tinacms/alerts
+
+
+
+
+
+## [0.1.6-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/alerts@0.1.5...@tinacms/alerts@0.1.6-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package @tinacms/alerts
+
+
+
+
+
 ## [0.1.5](https://github.com/tinacms/tinacms/compare/@tinacms/alerts@0.1.5-alpha.0...@tinacms/alerts@0.1.5) (2020-06-23)
 
 **Note:** Version bump only for package @tinacms/alerts