+++ packages/@tinacms/react-sidebar/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-sidebar@0.3.1-alpha.0...@tinacms/react-sidebar@0.3.1) (2020-07-07)
+
+**Note:** Version bump only for package @tinacms/react-sidebar
+
+
+
+
+
+## [0.3.1-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-sidebar@0.3.0...@tinacms/react-sidebar@0.3.1-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package @tinacms/react-sidebar
+
+
+
+
+
 # [0.3.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-sidebar@0.3.0-alpha.0...@tinacms/react-sidebar@0.3.0) (2020-06-29)
 
 **Note:** Version bump only for package @tinacms/react-sidebar