+++ packages/react-tinacms-date/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/tinacms/tinacms/compare/react-tinacms-date@0.1.1-alpha.0...react-tinacms-date@0.1.1) (2020-07-07)
+
+**Note:** Version bump only for package react-tinacms-date
+
+
+
+
+
+## [0.1.1-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms-date@0.1.0...react-tinacms-date@0.1.1-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package react-tinacms-date
+
+
+
+
+
 # [0.1.0](https://github.com/tinacms/tinacms/compare/react-tinacms-date@0.1.0-alpha.0...react-tinacms-date@0.1.0) (2020-06-29)
 
 **Note:** Version bump only for package react-tinacms-date