+++ packages/tinacms/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.22.1](https://github.com/tinacms/tinacms/compare/tinacms@0.22.1-alpha.0...tinacms@0.22.1) (2020-07-07)
+
+**Note:** Version bump only for package tinacms
+
+
+
+
+
+## [0.22.1-alpha.0](https://github.com/tinacms/tinacms/compare/tinacms@0.22.0...tinacms@0.22.1-alpha.0) (2020-07-04)
+
+**Note:** Version bump only for package tinacms
+
+
+
+
+
 # [0.22.0](https://github.com/tinacms/tinacms/compare/tinacms@0.22.0-alpha.0...tinacms@0.22.0) (2020-06-29)
 
 **Note:** Version bump only for package tinacms